Reload mechanics

+++ Assets/_Project/_Scripts/ScriptableObjects/PlayerData.cs
@@ -5,5 +5,5 @@ public class PlayerData : ScriptableObject
 {
     [field: SerializeField] public float GunDamage { get; set; }
     [field: SerializeField] public float GunReloadTime { get; set; }
-    [field: SerializeField] public float GunBulletAmount { get; set; }
+    [field: SerializeField] public int GunBulletAmount { get; set; }
 }
